requirements.txt
pyfiglet
pysocks
requests
